refactor - rename folders
