+++ task-2/libs/ddbClient.js
