+++ task-2/libs/ddbDocClient.js
